+++ app.js
@@ -4,34 +4,39 @@ const app = express();
 const cors = require("cors");
 const mongoose = require("mongoose");
 require("./config/joiObjectId")()
+const multer = require("multer")
+const upload = multer();
+require("express-async-errors");
 
 /* Routes */
 const usersRouter = require("./routes/users");
 const podcastsRouter = require("./routes/podcasts");
+const authorsRouter = require("./routes/authors");
 // const User = require('./models/user')
 
 
 /* Database connections */
-// mongoose.connect("mongodb://localhost:27017/pressTest",{ useNewUrlParser: true ,useUnifiedTopology: true })
-// .then(()=>{
-//     console.log('Connected to mongodb')
-// })
-// .catch((err)=> {
-//     return err.message
-// })
+mongoose.connect("mongodb://localhost:27017/pressPlay",{ useNewUrlParser: true ,useUnifiedTopology: true })
+.then(()=>{
+    console.log('Connected to mongodb')
+})
+.catch((err)=> {
+    return err.message
+})
 
-mongoose
-  .connect(process.env.MONGO_URI, {
-    useNewUrlParser: true,
-    useFindAndModify: false,
-    useCreateIndex: true,
-    useUnifiedTopology: true,
-  })
-  .then(() => "You are now connected to Mongo!")
-  .catch((err) => console.error("Something went wrong", err));
+// mongoose
+//   .connect(process.env.MONGO_URI, {
+//     useNewUrlParser: true,
+//     useFindAndModify: false,
+//     useCreateIndex: true,
+//     useUnifiedTopology: true,
+//   })
+//   .then(() => "You are now connected to Mongo!")
+//   .catch((err) => console.error("Something went wrong", err));
 
 //middlewares
-
+app.use(upload.any()); 
+app.use(express.urlencoded({extended:false})); 
 app.use(express.json());
 
 //to prevent cors errors
@@ -41,9 +46,11 @@ app.get("/", (req, res, next) => {
   res.send(`Welcome to Press Play API`);
 });
 
-app.use("/users", usersRouter);
+app.use("/api/users", usersRouter);
 
 app.use("/api/podcasts", podcastsRouter);
+
+app.use("/api/authors", authorsRouter);
 
 app.use((req, res, next) => {
   res.status(404).send(new Error());